# api/http/openapi.yaml
openapi: 3.1.0
info:
  title: MatchLab Gateway API
  version: 2.3.0
  description: |
    Текущий публичный HTTP API MatchLab: учётные записи, сеансы, подтверждение почты и файлы.
    Это описание реализованных маршрутов Gateway. Для локальной проверки сначала
    зарегистрируйтесь, подтвердите email из письма, затем войдите и укажите полученный токен доступа в Bearer Auth.
    Для проверки почты сначала добавьте разрешённые домены HSE локальным seed-скриптом.
    Письмо с подтверждением появится в Mailpit по адресу http://localhost:8025.
    Токен из ссылки в письме передайте в метод подтверждения адреса.
servers:
  - url: http://localhost:8080
    description: Gateway в локальном Docker Compose
...
paths:
  /api/v1/media/files:
    post:
      tags: [Файлы]
      summary: Подготовить загрузку файла
      operationId: createMediaUpload
      description: |
        Требует сеанс и подтверждённую почту. Владелец определяется из сеанса.
        Возвращает pending и подписанный PUT. Передайте байты файла напрямую по upload_url,
        с выданными headers, затем вызовите complete. Не используйте multipart/form-data.
        В браузере Content-Length формируется автоматически. is_public по умолчанию false.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMediaUpload'
            example:
              original_name: avatar.png
              content_type: image/png
              size_bytes: 1024
              is_public: true
      responses:
        '201':
          description: Создана запись pending. Файл ещё не загружен.
          headers:
            Location:
              schema: {type: string}
              description: HTTP-путь к метаданным файла.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaUpload'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/MediaForbidden'
        '409':
          $ref: '#/components/responses/MediaConflict'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
        '504':
          $ref: '#/components/responses/MediaTimeout'
  /api/v1/media/files/{id}/complete:
    parameters:
      - $ref: '#/components/parameters/MediaFileID'
    post:
      tags: [Файлы]
      summary: Подтвердить загрузку файла
      operationId: completeMediaUpload
      description: |
        Только владелец с подтверждённой почтой, включая публичные файлы.
        Тело запроса не требуется. Media проверяет объект в S3 и переводит pending в ready.
        Проверяются тип, лимит и ожидаемый размер (для новых загрузок).
        Pending без обновления БД 24 часа также обрабатывает фоновая задача:
        подтверждает корректный объект или очищает S3 и сохраняет failed.
        Повтор возвращает сохранённые метаданные ready.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Файл подтверждён.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaFileResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/MediaForbidden'
        '404':
          $ref: '#/components/responses/MediaNotFound'
        '409':
          $ref: '#/components/responses/MediaConflict'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
        '504':
          $ref: '#/components/responses/MediaTimeout'
  /api/v1/media/files/{id}:
    parameters:
      - $ref: '#/components/parameters/MediaFileID'
    delete:
      tags: [Файлы]
      summary: Удалить файл из S3
      operationId: deleteMediaFile
      description: |
        Только владелец с действующим сеансом и подтверждённой почтой, включая публичные файлы.
        user_id берётся из сеанса. Чужой или отсутствующий файл возвращает 404.
        Для ready удаляет объект из S3 и сохраняет deleted в БД; повтор владельцем возвращает 204.
        Если удаление не завершилось, файл остаётся deleting, скрывается из чтения и
        обрабатывается фоновой задачей при старте Media и каждые 10 минут.
        pending и failed возвращают 409; для отмены multipart используйте DELETE /{id}/multipart.
        Удаление объекта не отзывает ранее выданные подписанные PUT-ссылки.
      security: [ { bearerAuth: [ ] } ]
      responses:
        '204': { description: Объект удалён или файл уже был удалён этим владельцем. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/MediaForbidden' }
        '404': { $ref: '#/components/responses/MediaNotFound' }
        '409': { $ref: '#/components/responses/MediaConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '504': { $ref: '#/components/responses/MediaTimeout' }
    get:
      tags: [Файлы]
      summary: Получить метаданные файла
      operationId: getMediaFile
      description: |
        Публичный файл доступен без сеанса; приватный — только владельцу.
        Если Authorization передан, он проверяется вместе с подтверждением почты.
        Недействительный токен возвращает 401, а не гостевой доступ.
        Не передавайте user_id в запросе. Ответы не кешируются (Cache-Control no-store).
      security:
        - { }
        - bearerAuth: [ ]
      responses:
        '200':
          description: Метаданные без ссылки на содержимое; доступны и в состоянии pending.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaFileResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/MediaForbidden'
        '404':
          $ref: '#/components/responses/MediaNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
        '504':
          $ref: '#/components/responses/MediaTimeout'
  /api/v1/media/files/{id}/download-url:
    parameters:
      - $ref: '#/components/parameters/MediaFileID'
    get:
      tags: [Файлы]
      summary: Получить временную ссылку на скачивание
      operationId: createMediaDownloadURL
      description: |
        Для ready. Публичный файл доступен гостю, приватный — владельцу.
        Переданный Authorization проверяется вместе с подтверждением почты.
        Ответ Cache-Control no-store. Выполните GET по download_url напрямую в S3,
        без Bearer-токена Gateway. Ссылка даёт доступ любому её получателю до истечения TTL.
      security:
        - { }
        - bearerAuth: [ ]
      responses:
        '200':
          description: Подписанная GET-ссылка на содержимое.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaDownload'
              example:
                download_url: 'http://localhost:8333/matchlab-media/users/example/file?X-Amz-Signature=example'
                expires_at_unix: 1791543600
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/MediaForbidden'
        '404':
          $ref: '#/components/responses/MediaNotFound'
        '409':
          $ref: '#/components/responses/MediaConflict'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
        '504':
          $ref: '#/components/responses/MediaTimeout'
  /api/v1/media/files/multipart:
    post:
      tags: [Файлы]
      summary: Начать загрузку файла частями
      operationId: createMediaMultipart
      description: |
        Требует сеанс и подтверждённую почту. Возвращает pending-файл и параметры сессии.
        Размер части задаёт Media; последняя часть может быть меньше. Байты передаются
        напрямую в S3 через part-urls. Существующая загрузка одним PUT также доступна.
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateMediaUpload'}
            example: {original_name: report.pdf, content_type: application/pdf, size_bytes: 8388611, is_public: false}
      responses:
        '201':
          description: Созданы pending-файл и сессия загрузки.
          headers:
            Location:
              schema: {type: string}
              description: Путь к состоянию multipart-сессии.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MediaMultipartState'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/MediaForbidden'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
        '504': {$ref: '#/components/responses/MediaTimeout'}
  /api/v1/media/files/{id}/multipart:
    parameters:
      - $ref: '#/components/parameters/MediaFileID'
    get:
      tags: [Файлы]
      summary: Получить прогресс загрузки частями
      operationId: getMediaMultipart
      description: |
        Только владелец с подтверждённой почтой, даже для публичного файла.
        Прогресс учитывает полностью принятые S3 части. В SeaweedFS после повторного PUT
        возможны несколько ETag одного номера; uploaded_bytes учитывает этот номер один раз.
        Сохраняйте ETag ответов PUT для завершения. Ответ Cache-Control no-store.
      security: [ { bearerAuth: [ ] } ]
      responses:
        '200':
          description: Состояние сессии и принятые части; для completed uploaded_bytes равен размеру файла.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MediaMultipartState' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/MediaForbidden' }
        '404': { $ref: '#/components/responses/MediaNotFound' }
        '409': { $ref: '#/components/responses/MediaConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '504': { $ref: '#/components/responses/MediaTimeout' }
    delete:
      tags: [Файлы]
      summary: Отменить загрузку частями
      operationId: abortMediaMultipart
      description: |
        Только владелец с подтверждённой почтой. Прерывает S3-сессию и переводит файл
        в failed. Повтор отмены возвращает 204. Для completing и completed возвращает 409.
        Перед отменой остановите отправку частей на клиенте.
      security: [ { bearerAuth: [ ] } ]
      responses:
        '204': { description: Загрузка отменена. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/MediaForbidden' }
        '404': { $ref: '#/components/responses/MediaNotFound' }
        '409': { $ref: '#/components/responses/MediaConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '504': { $ref: '#/components/responses/MediaTimeout' }
  /api/v1/media/files/{id}/multipart/part-urls:
    parameters:
      - $ref: '#/components/parameters/MediaFileID'
    post:
      tags: [Файлы]
      summary: Получить ссылки для отправки частей
      operationId: createMediaPartURLs
      description: |
        Только владелец с подтверждённой почтой. Допускает 1–100 уникальных номеров
        от 1 до part_count. Передайте каждую часть напрямую в S3 методом PUT с выданными
        headers и ровно size_bytes байтов. Не используйте multipart/form-data.
        Сохраните ETag ответа PUT. Для повтора части или истечения ссылки запросите новую.
        Ссылка действует не дольше срока сессии. Completing, completed, aborted и просроченная сессия дают 409.
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [part_numbers]
              properties:
                part_numbers:
                  type: array
                  minItems: 1
                  maxItems: 100
                  uniqueItems: true
                  items: {type: integer, format: int32, minimum: 1, maximum: 10000}
            example: {part_numbers: [1, 2]}
      responses:
        '200':
          description: Подписанные PUT-ссылки для частей.
          content:
            application/json:
              schema:
                type: object
                required: [parts]
                properties:
                  parts:
                    type: array
                    items: {$ref: '#/components/schemas/MediaPartURL'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/MediaForbidden'}
        '404': {$ref: '#/components/responses/MediaNotFound'}
        '409': {$ref: '#/components/responses/MediaConflict'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
        '504': {$ref: '#/components/responses/MediaTimeout'}
  /api/v1/media/files/{id}/multipart/complete:
    parameters:
      - $ref: '#/components/parameters/MediaFileID'
    post:
      tags: [Файлы]
      summary: Собрать и подтвердить загруженные части
      operationId: completeMediaMultipart
      description: |
        Только владелец с подтверждённой почтой. Передайте полный список номеров и ETag
        из ответов PUT, включая кавычки ETag. Порядок списка произвольный, номера уникальны.
        Media проверяет части и размер, собирает объект в S3 и переводит файл в ready.
        Лимит JSON — 1 МиБ. Повтор с тем же списком безопасен после потери ответа.
        Прерванное completing восстанавливается Media. При конфликте ETag загрузка
        возвращается в uploading: получите состояние, повторите части и завершение.
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [parts]
              properties:
                parts:
                  type: array
                  minItems: 1
                  maxItems: 10000
                  items:
                    type: object
                    additionalProperties: false
                    required: [part_number, etag]
                    properties:
                      part_number: {type: integer, format: int32, minimum: 1, maximum: 10000}
                      etag: {type: string, minLength: 1}
            example:
              parts:
                - {part_number: 1, etag: '"ETAG_FROM_FIRST_PUT"'}
                - {part_number: 2, etag: '"ETAG_FROM_SECOND_PUT"'}
      responses:
        '200':
          description: Файл готов; повтор возвращает сохранённые метаданные.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MediaFileResult'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/MediaForbidden'}
        '404': {$ref: '#/components/responses/MediaNotFound'}
        '409': {$ref: '#/components/responses/MediaConflict'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
        '504': {$ref: '#/components/responses/MediaTimeout'}
...
components:
  parameters:
    MediaFileID:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Токен access_token из ответа POST /api/v1/auth/login.
  schemas:
    MediaMultipartState:
      type: object
      required: [file, status, expected_size_bytes, part_size_bytes, part_count, expires_at_unix, uploaded_bytes, parts]
      properties:
        file: {$ref: '#/components/schemas/MediaFile'}
        status: {type: string, enum: [uploading, completing, completed, aborted]}
        expected_size_bytes: {type: integer, format: int64, minimum: 1}
        part_size_bytes: {type: integer, format: int64, minimum: 5242880, maximum: 5368709120}
        part_count: {type: integer, format: int32, minimum: 1, maximum: 10000}
        expires_at_unix: {type: integer, format: int64}
        uploaded_bytes: {type: integer, format: int64, minimum: 0}
        parts:
          type: array
          items:
            type: object
            required: [part_number, etag, size_bytes]
            properties:
              part_number: {type: integer, format: int32}
              etag: {type: string}
              size_bytes: {type: integer, format: int64}
    MediaPartURL:
      type: object
      required: [part_number, size_bytes, upload_url, method, headers, expires_at_unix]
      properties:
        part_number: {type: integer, format: int32}
        size_bytes: {type: integer, format: int64}
        upload_url: {type: string, format: uri}
        method: {type: string, enum: [PUT]}
        headers:
          type: object
          additionalProperties: {type: string}
        expires_at_unix: {type: integer, format: int64}
    CreateMediaUpload:
      type: object
      additionalProperties: false
      required: [original_name, content_type, size_bytes]
      properties:
        original_name:
          type: string
          minLength: 1
          maxLength: 255
          description: Имя без пути и управляющих символов.
        content_type:
          type: string
          description: Разрешённый MIME-тип без параметров; список задаётся настройками Media.
          examples: [application/pdf, image/jpeg, image/png]
        size_bytes:
          type: integer
          format: int64
          minimum: 1
          description: Размер в байтах; лимит из конфигурации Media (локально 100 MiB).
        is_public:
          type: boolean
          default: false
    MediaFile:
      type: object
      required: [id, owner_user_id, original_name, content_type, status, created_at_unix, is_public]
      properties:
        id: {type: string, format: uuid}
        owner_user_id: {type: string, format: uuid}
        original_name: {type: string}
        content_type: {type: string}
        size_bytes:
          type: integer
          format: int64
          description: Фактический размер; до подтверждения поле отсутствует.
        status:
          type: string
          enum: [pending, ready, deleting, deleted, failed]
        created_at_unix: {type: integer, format: int64}
        uploaded_at_unix:
          type: integer
          format: int64
          description: Время подтверждения в секундах Unix; до подтверждения отсутствует.
        is_public: {type: boolean}
    MediaFileResult:
      type: object
      required: [file]
      properties:
        file:
          $ref: '#/components/schemas/MediaFile'
    MediaUpload:
      type: object
      required: [file, upload_url, method, headers, expires_at_unix]
      properties:
        file:
          $ref: '#/components/schemas/MediaFile'
        upload_url: {type: string, format: uri}
        method: {type: string, enum: [PUT]}
        headers:
          type: object
          additionalProperties: {type: string}
          description: Заголовки подписанного PUT. Content-Length в браузере задаёт сам браузер.
        expires_at_unix: {type: integer, format: int64}
    MediaDownload:
      type: object
      required: [download_url, expires_at_unix]
      properties:
        download_url: {type: string, format: uri}
        expires_at_unix: {type: integer, format: int64}
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
  responses:
    MediaNotFound:
      description: Файл отсутствует, удалён или недоступен читателю. Подтверждать чужой публичный файл также нельзя.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    MediaConflict:
      description: Неподходящее состояние, объект ещё не загружен, неверные метаданные объекта или дубликат.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    MediaForbidden:
      description: Подтвердите почту или доступ запрещён.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    MediaTimeout:
      description: Превышено время ожидания ответа Media.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Внутренняя ошибка без подробностей реализации.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Некорректный JSON или значения запроса.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: некорректное тело запроса
    Unauthorized:
      description: Токен доступа отсутствует или недействителен.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: недействительный сеанс
...
    Unavailable:
      description: Зависимый сервис или SMTP-сервер недоступен.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: сервис недоступен